+++ jvue-front/nuxt.config.js
@@ -62,7 +62,7 @@ module.exports = {
   axios: {
     // See https://github.com/nuxt-community/axios-module#options
     baseURL: development
-      ? "http://localhost:8001/api"
+      ? "http://localhost:8002/api"
       : "http://v4.terwergreen.com:8002/api"
   },
