+++ jvue-front/pages/s-v3/_keyword.vue
@@ -1,20 +1,20 @@
 <template>
   <el-container>
-    <el-main>
+    <el-main class="el-container-dark">
       <el-row>
         <el-col :xs="0" :md="2">&nbsp;</el-col>
         <el-col :xs="24" :md="20">
-          <el-main>
+          <el-main class="el-main-dark">
             <el-container>
-              <el-main>
+              <el-main class="el-main-dark">
                 <el-container>
-                  <el-header>
+                  <el-header class="el-header-dark">
                     <HeaderTime />
                   </el-header>
-                  <el-header>
+                  <el-header class="el-header-dark">
                     <Header />
                   </el-header>
-                  <el-main>
+                  <el-main class="el-main-dark">
                     <Body :keyword="keyword" :post-list="postListArray" />
                   </el-main>
                 </el-container>
@@ -38,13 +38,12 @@
 
 <script>
 import { getLogger } from "../../util/logger";
-import HeaderTime from "../../components/themes/default/HeaderTime";
-import Header from "../../components/themes/default/Header";
-import Body from "../../components/themes/default/Body";
-import Footer from "../../components/themes/default/Footer";
-import FriendLink from "../../components/themes/default/FriendLink";
-
-const logger = getLogger("pages/post");
+import HeaderTime from "../../components/themes/dark/HeaderTime";
+import Header from "../../components/themes/dark/Header";
+import Body from "../../components/themes/dark/Body";
+import Footer from "../../components/themes/dark/Footer";
+import FriendLink from "../../components/themes/dark/FriendLink";
+const logger = getLogger("pages/index");
 
 export default {
   components: { HeaderTime, Header, Body, Footer, FriendLink },
@@ -52,7 +51,7 @@ export default {
     const siteConfigResult = await $axios.$post("/site/config/list");
     const siteConfigObj =
       siteConfigResult.status === 1 ? siteConfigResult.data : {};
-    logger.info("fetch siteConfig and postList finish");
+    logger.info("fetch siteConfig finish");
     return { siteConfigObj };
   },
   data() {
@@ -92,7 +91,7 @@ export default {
 
 <style lang="scss">
 @import "../common.css";
-@import "../default.css";
+@import "../dark.css";
 </style>
 
 <style scoped></style>
